renamed login folder to auth
