+++ client/src/app/app.routes.ts
@@ -1,6 +1,5 @@
-import { AuthenticationService } from './login/services/authentication.service';
 import { Routes } from '@angular/router';
-import { AuthComponent } from './login/auth/auth.component';
+import { AuthComponent } from './auth/auth/auth.component';
 import { HomeComponent } from './home/home.component';
 import { isUserLoggedInGuard } from './routeGuards/isUserLoggedIn.guard';
