+++ client/src/app/routeGuards/isUserLoggedIn.guard.ts
@@ -1,5 +1,5 @@
 import { CanActivateFn } from '@angular/router';
-import { AuthenticationService } from '../login/services/authentication.service';
+import { AuthenticationService } from '../auth/services/authentication.service';
 import { inject } from '@angular/core';
 
 export const isUserLoggedInGuard: CanActivateFn = (route, state) => {
